Fix a few implicit casts flagged by CodeQL

+++ src/main/java/uk/org/okapibarcode/backend/CodablockF.java
@@ -833,7 +833,7 @@ public class CodablockF extends Symbol {
                 } else {
                     black = true;
                 }
-                x += (double)(pattern[yBlock].charAt(xBlock) - '0');
+                x += pattern[yBlock].charAt(xBlock) - '0';
             }
             y += h;
             if (y > symbol_height) {
